requirements.txt
flask>=3.0.0
openai>=1.0.0
sounddevice>=0.4.6
soundfile>=0.12.1
openai-whisper>=20231117
numpy>=1.24.0